requirements.txt
Pillow
blobfile
DALL-E
ftfy
kornia
mypy
matplotlib
numpy
opencv-python
pytest
requests
regex
scipy
torch
torchvision
tensorflow
ffmpeg
SpeechRecognition
googletrans